// msquare/lms/gurukul/src/app/core/models/auth.models.ts
export class User {
  id: number = 0;
  password: string = '';
  photo: string = '';
  email: string = '';
  role: String = '';
  accessToken: string = '';
  refreshToken: string = '';
  firstName: string = '';
  lastName: string = '';
  salutation: string = '';
  dob: string = '';
  lastLoginOn: string = '';
  rememberMe: boolean = false;
  mobile: number = 0;
  whatsappNo: number=0;
  language: string = '';
  isAgree: boolean = false;
}